spline #971 refactoring: controller code deduplication
