+++ consumer-services/src/main/scala/za/co/absa/spline/consumer/service/repo/DataSourceRepository.scala
@@ -19,27 +19,7 @@ import za.co.absa.spline.consumer.service.model._
 
 import scala.concurrent.{ExecutionContext, Future}
 
-trait DataSourceRepository {
-
-  def getTimestampRange(
-    asAtTime: Long,
-    maybeSearchTerm: Option[String],
-    maybeAppend: Option[Boolean],
-    maybeApplicationId: Option[String],
-    maybeDataSourceUri: Option[String])
-    (implicit ec: ExecutionContext): Future[(Long, Long)]
-
-  def find(
-    asAtTime: Long,
-    maybeWriteTimestampStart: Option[Long],
-    maybeWriteTimestampEnd: Option[Long],
-    pageRequest: PageRequest,
-    sortRequest: SortRequest,
-    maybeSearchTerm: Option[String],
-    maybeAppend: Option[Boolean],
-    maybeWriteApplicationId: Option[String],
-    maybeDataSourceUri: Option[String])
-    (implicit ec: ExecutionContext): Future[(Seq[WriteEventInfo], Long)]
+trait DataSourceRepository extends AbstractExecutionEventRepository {
 
   def findByUsage(
     execPlanId: ExecutionPlanInfo.Id,
